Added basic stack spilling and floating point support
Stack spilling now works in a basic way. If we cannot find a valid type
and size match in the register table, we automatically spill to the
stack. Also, floating point fastcall registers now correctly get
allocated. Stack spilling behavior for integers and floating point types
was verified to be the same as llvm; namely, arguments are spilled to
the stack in the order they are encountered while iterating from left
to right.