Added basic return value deduction and cleaned up
Added a simple return value deduction scheme that iterates through all
of the return instructions in a function and gets their type. As of now
it assigns all return values to RAX as per x64-SysV calling convention
but will add a more general approach later. Also removed the
AllocationBinding struct because it is no longer needed. It is rendered
obsolete by the previous migration to ValueDecl and ParameterDecl.