Fixed struct index handling bug
Fixed a bug which caused incorrect bitcode when accessing any element of
a struct besides the first. Fixing the vector in GetIndexesInto to be
pass by reference allowed changes to indexes to persist outside of the
function. This allowed getelementptr instructions with multiple
arguments be created. Changing the initial base_index to be of i32_type
is an artifact of an llvm source-documented constraint. Apparently,
all indexes to structure types must be integers or integer vectors
that are 32-bits wide.